# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AiError {
  error: string;
}

export interface AiChatRequest {
  /**
   * @minLength 1
   * @maxLength 8000
   */
  message: string;
}

export interface AiChatReply {
  reply: string;
}

export interface AiSummarizeTextRequest {
  /**
   * @minLength 1
   * @maxLength 60000
   */
  text: string;
}

export interface AiSummarizeUrlRequest {
  /**
   * @minLength 8
   * @maxLength 2048
   */
  url: string;
}

export interface AiSummarizeVideoRequest {
  /**
   * @minLength 1
   * @maxLength 60000
   */
  transcript: string;
}

export interface AiSummaryResponse {
  summary: string;
}

export interface AiTranslateRequest {
  /**
   * @minLength 1
   * @maxLength 8000
   */
  text: string;
  /**
   * @minLength 1
   * @maxLength 64
   */
  targetLanguage: string;
}

export interface AiTranslationResult {
  translation: string;
}

/**
 * Hint of the audio container format.
 */
export type AiTranscribeRequestFormat =
  (typeof AiTranscribeRequestFormat)[keyof typeof AiTranscribeRequestFormat];

export const AiTranscribeRequestFormat = {
  wav: "wav",
  mp3: "mp3",
  m4a: "m4a",
  webm: "webm",
  ogg: "ogg",
  aac: "aac",
  flac: "flac",
  mp4: "mp4",
} as const;

export interface AiTranscribeRequest {
  /**
   * Base64-encoded audio bytes (no data URL prefix).
   * @minLength 1
   */
  audioBase64: string;
  /** Hint of the audio container format. */
  format?: AiTranscribeRequestFormat;
}

export interface AiTranscriptionResult {
  text: string;
}